Update how data is accessed
In SvleteKit, data is returned through load functions. See https://kit.svelte.dev/docs/load

The load function returns the data, which is accessed by pages. In our case, because the sitename needed to be accessed in +layout.svelte, it has to be returned in +layout.server.js or +layout.js